Increases window of how far back we will go to fetch orders [Ref #160]
This is a temporary "fix" for #160 where we occasionally fail to correctly
fetch orders. In both cases where this happened, increasing the sync window
seemed to recover the missing orders.

My guess is the underlying issue is a timezone mismatch or something related to
us not fetching far back enough if we've missed a few fetches.